fix PartSource1D_Load for nq kind

+++ src/PartSource/PartSource1D_Load.h
@@ -54,7 +54,7 @@ public:
     int loadStep;
     // Number of particles loaded in one cell at one loadStep
     int loadNumber;
-    double loadNumber_heat;
+    int loadNumber_heat;
     // loadRem = loadStep * loadDn *... - loadNumber
     double loadRem;
     double loadRemTot;
